+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/NetworkRequest.kt
@@ -1,9 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
-
-data class NetworkRequest(
-  val url: String,
-  val body: String? = null,
-  val error: String? = null,
-)
-