+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/RemoteRepo_ListParams.kt
@@ -1,9 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
-
-data class RemoteRepo_ListParams(
-  val query: String? = null,
-  val first: Int,
-  val afterId: String? = null,
-)
-