+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/WriteFileOptions.kt
@@ -1,8 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
-
-data class WriteFileOptions(
-  val overwrite: Boolean? = null,
-  val ignoreIfExists: Boolean? = null,
-)
-