+++ vscode/src/testutils/mocks.ts
@@ -17,7 +17,9 @@ import {
     ps,
 } from '@sourcegraph/cody-shared'
 
+import path from 'node:path'
 import { AgentEventEmitter as EventEmitter } from './AgentEventEmitter'
+import { AgentWorkspaceEdit as WorkspaceEdit } from './AgentWorkspaceEdit'
 import { Uri } from './uri'
 
 export { Uri } from './uri'
@@ -25,7 +27,6 @@ export { Uri } from './uri'
 export { AgentEventEmitter as EventEmitter } from './AgentEventEmitter'
 export { AgentWorkspaceEdit as WorkspaceEdit } from './AgentWorkspaceEdit'
 export { Disposable } from './Disposable'
-import { AgentWorkspaceEdit as WorkspaceEdit } from './AgentWorkspaceEdit'
 
 /**
  * This module defines shared VSCode mocks for use in every Vitest test.
@@ -548,6 +549,7 @@ export const workspaceFs: typeof vscode_types.workspace.fs = {
         }
     },
     writeFile: async (uri, content) => {
+        await fspromises.mkdir(path.dirname(uri.fsPath), { recursive: true })
         await fspromises.writeFile(uri.fsPath, content)
     },
     delete: async (uri, options) => {